+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # (Unreleased; add upcoming change notes here)
 
+- fix bug where we should incorrect revision save times in history viewer,
+  also show revision save time down to the second level
+
 # 0.3.0 (2019-03-06)
 
 - don't show "new notebook" button on user pages that don't belong to