+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # (Unreleased; add upcoming change notes here)
 
+# 0.4.0 (2019-03-09)
+
 - iodide notebooks now automatically save to the server
 - fix bug where we showed incorrect revision save times in history viewer,
   also show revision save time down to the second level